Refactor score package
