+++ discgolfbot/score/udisc/udisc_csv_types.py
